Rename Navigator to NavigatorSearch
