--- src/components/Navigator/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Navigator";
-export * from "./Navigator";
