+++ src/components/NavigatorSearch/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./NavigatorSearch";
+export * from "./NavigatorSearch";
