+++ src/components/NavigatorSearch/modes/catalog.ts
