+++ src/components/NavigatorSearch/modes/commands/index.ts
