+++ src/components/NavigatorSearch/modes/default/views.ts
