+++ src/components/NavigatorSearch/modes/index.ts
