+++ src/components/NavigatorSearch/queries/useCatalogSearch.ts
