+++ src/components/NavigatorSearch/useQuickSearch.ts
